requirements.txt
feedparser
google-generativeai
atproto
python-dotenv
